requirements.txt
Flask
Flask-SQLAlchemy
pymysql
boto3
flask-cors